worker_rpa/requirements.txt
# Worker RPA
requests
pywin32
python-dotenv
sqlalchemy
psycopg2-binary
